(refactor) : code refactor of department filter
