+++ src/routes/departmentRoutes.ts
@@ -5,8 +5,7 @@ const departmentController = new DepartmentController();
 export const departmentRouter = Router();
 
 departmentRouter.post('/', departmentController.createDepartment);
-departmentRouter.get('/', departmentController.getDepartmentsFilter);
-// departmentRouter.get('/', departmentController.getDepartments);
+departmentRouter.get('/', departmentController.getDepartmentsByFilter);
 departmentRouter.get('/:id', departmentController.getDepartmentById);
 departmentRouter.put('/:id', departmentController.updateDepartment);
 departmentRouter.delete('/:id', departmentController.deleteDepartment);
